Modularity
of frontend injection code and styles
