+++ Web/Injection/saves.css
@@ -1,0 +1,187 @@
+.je-saves-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
+    gap: 18px;
+    padding: 24px;
+}
+.je-save-card {
+    background: rgba(255,255,255,0.05);
+    border: 1px solid rgba(255,255,255,0.08);
+    border-radius: 10px;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+    transition: transform 0.15s ease, border-color 0.15s ease;
+    cursor: default;
+}
+.je-save-card:hover {
+    transform: translateY(-3px);
+    border-color: rgba(0,164,220,0.5);
+}
+.je-save-art {
+    width: 100%;
+    aspect-ratio: 16/9;
+    object-fit: cover;
+    background: rgba(0,0,0,0.4);
+    display: block;
+    flex-shrink: 0;
+}
+.je-save-art-poster {
+    width: 100%;
+    aspect-ratio: 2/3;
+    object-fit: cover;
+    background: rgba(0,0,0,0.4);
+    display: block;
+    flex-shrink: 0;
+}
+.je-save-art-placeholder {
+    width: 100%;
+    aspect-ratio: 16/9;
+    background: rgba(0,0,0,0.35);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-shrink: 0;
+}
+.je-save-art-placeholder .material-icons {
+    font-size: 56px;
+    color: rgba(255,255,255,0.15);
+}
+.je-save-body {
+    padding: 12px 14px 14px;
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+    flex: 1;
+}
+.je-save-title {
+    font-size: 0.88rem;
+    font-weight: 600;
+    color: #fff;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    line-height: 1.3;
+}
+.je-save-badges {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 4px;
+}
+.je-save-badge {
+    font-size: 10px;
+    font-weight: 700;
+    letter-spacing: .03em;
+    padding: 2px 6px;
+    border-radius: 4px;
+    line-height: 1.4;
+}
+.je-save-badge-platform {
+    background: rgba(255,255,255,0.1);
+    color: #ccc;
+    border: 1px solid rgba(255,255,255,0.15);
+}
+.je-save-badge-region {
+    background: rgba(0,164,220,0.8);
+    color: #fff;
+}
+.je-save-badge-disc {
+    background: rgba(220,140,0,0.85);
+    color: #fff;
+}
+.je-save-badge-slot {
+    background: rgba(82,181,75,0.25);
+    color: #7ed67a;
+    border: 1px solid rgba(82,181,75,0.35);
+}
+.je-save-meta {
+    font-size: 0.75rem;
+    color: rgba(255,255,255,0.45);
+    line-height: 1.4;
+}
+.je-save-actions {
+    display: flex;
+    gap: 8px;
+    margin-top: auto;
+    padding-top: 10px;
+}
+.je-save-btn {
+    flex: 1;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 5px;
+    padding: 7px 10px;
+    border-radius: 6px;
+    font-size: 0.78rem;
+    font-weight: 600;
+    cursor: pointer;
+    border: none;
+    transition: background 0.15s ease, opacity 0.15s ease;
+    text-decoration: none;
+}
+.je-save-btn .material-icons { font-size: 15px; }
+.je-save-btn-play {
+    background: rgba(0,164,220,0.85);
+    color: #fff;
+}
+.je-save-btn-play:hover { background: rgba(0,164,220,1); }
+.je-save-btn-delete {
+    background: rgba(220, 53, 69, 0.85);
+    color: #fff;
+}
+.je-save-btn-delete:hover { 
+    background: rgba(220, 53, 69, 1); 
+}
+.je-save-btn-dl {
+    background: rgba(255,255,255,0.08);
+    color: rgba(255,255,255,0.75);
+    border: 1px solid rgba(255,255,255,0.12);
+}
+.je-save-btn-dl:hover { background: rgba(255,255,255,0.15); }
+.je-save-btn-romm-push {
+    background: rgba(82,181,75,0.15);
+    color: #52B54B;
+    border: 1px solid rgba(82,181,75,0.3);
+}
+.je-save-btn-romm-push:hover { background: rgba(82,181,75,0.28); }
+.je-save-btn-romm-pull {
+    background: rgba(0,164,220,0.12);
+    color: #00a4dc;
+    border: 1px solid rgba(0,164,220,0.25);
+}
+.je-save-btn-romm-pull:hover { background: rgba(0,164,220,0.25); }
+.je-saves-empty {
+    text-align: center;
+    color: rgba(255,255,255,0.35);
+    padding: 80px 24px;
+    font-size: 1rem;
+}
+.je-saves-empty .material-icons { font-size: 64px; display: block; margin-bottom: 16px; opacity: 0.3; }
+.je-saves-header {
+    display: flex;
+    align-items: center;
+    gap: 16px;
+    padding: 20px 24px 4px;
+}
+.je-saves-filter {
+    background: rgba(255,255,255,0.07);
+    border: 1px solid rgba(255,255,255,0.12);
+    border-radius: 6px;
+    color: #fff;
+    padding: 6px 12px;
+    font-size: 0.82rem;
+    cursor: pointer;
+    outline: none;
+    transition: border-color 0.15s;
+}
+.je-saves-filter option {
+    background: #1a1a2e;
+    color: #fff;
+}
+.je-saves-filter:focus { border-color: #00a4dc; }
+.je-saves-count {
+    font-size: 0.82rem;
+    color: rgba(255,255,255,0.4);
+    margin-left: auto;
+}
